+++ chapter5/calc-expression/expr.h
@@ -21,17 +21,17 @@ typedef enum {
 
 struct expr {
        expr_t kind;
-       double value;
+       int value;
        struct expr *left;
        struct expr *right;
 };
 
 struct expr * expr_create( expr_t kind, struct expr *left, struct expr *right );
-struct expr * expr_create_value( double value );
+struct expr * expr_create_value( int value );
 
 void          expr_print( struct expr *e );
 void          expr_delete( struct expr *e );
-double        expr_evaluate( struct expr *e );
+int           expr_evaluate( struct expr *e );
 
 #endif
